Complete View

+++ src/Datiss.Budget.Services/CostCurrentConsumableService.cs
@@ -245,7 +245,7 @@ namespace Datiss.Budget.Services
             result.Add(new CalculationItemData
             {
                 Key = "CostCurrentConsumable_Cal1",
-                Value = await _uow.ExecuteScalar<int>(
+                Value = await _uow.ExecuteScalar<long>(
                                     "[dbo].[CostCurrentConsumable_Cal1] @YearId, @OrganizationId",
                                     parameters: sqlParams.ToArray())
             });
@@ -253,7 +253,7 @@ namespace Datiss.Budget.Services
             result.Add(new CalculationItemData
             {
                 Key = "CostCurrentConsumable_Cal2",
-                Value = await _uow.ExecuteScalar<int>(
+                Value = await _uow.ExecuteScalar<long>(
                                     "[dbo].[CostCurrentConsumable_Cal2] @YearId, @OrganizationId",
                                     parameters: sqlParams.ToArray())
             });
@@ -261,7 +261,7 @@ namespace Datiss.Budget.Services
             result.Add(new CalculationItemData
             {
                 Key = "CostCurrentConsumable_Cal3",
-                Value = await _uow.ExecuteScalar<int>(
+                Value = await _uow.ExecuteScalar<long>(
                          "[dbo].[CostCurrentConsumable_Cal3] @YearId, @OrganizationId",
                          parameters: sqlParams.ToArray())
             });
@@ -269,7 +269,7 @@ namespace Datiss.Budget.Services
             result.Add(new CalculationItemData
             {
                 Key = "CostCurrentConsumable_Cal4",
-                Value = await _uow.ExecuteScalar<int>(
+                Value = await _uow.ExecuteScalar<long>(
                          "[dbo].[CostCurrentConsumable_Cal4] @YearId, @OrganizationId",
                          parameters: sqlParams.ToArray())
             });
@@ -277,7 +277,7 @@ namespace Datiss.Budget.Services
             result.Add(new CalculationItemData
             {
                 Key = "CostCurrentConsumable_Cal5",
-                Value = await _uow.ExecuteScalar<int>(
+                Value = await _uow.ExecuteScalar<long>(
                          "[dbo].[CostCurrentConsumable_Cal5] @YearId, @OrganizationId",
                          parameters: sqlParams.ToArray())
             });
@@ -285,7 +285,7 @@ namespace Datiss.Budget.Services
             result.Add(new CalculationItemData
             {
                 Key = "CostCurrentConsumable_Cal6",
-                Value = await _uow.ExecuteScalar<int>(
+                Value = await _uow.ExecuteScalar<long>(
              "[dbo].[CostCurrentConsumable_Cal6] @YearId, @OrganizationId",
              parameters: sqlParams.ToArray())
             });
